Split out Map and LocationMarker components out of the map page
