+++ src/components/map/LocationMarker.test.js
@@ -1,0 +1,8 @@
+import React from "react";
+import { shallow } from "enzyme";
+import LocationMarker from "./LocationMarker";
+
+it("renders without crashing", () => {
+  const wrapper = shallow(<LocationMarker />);
+  expect(wrapper).toHaveLength(1);
+});
