+++ src/components/map/Map.test.js
@@ -1,0 +1,8 @@
+import React from "react";
+import { shallow } from "enzyme";
+import Map from "./Map";
+
+it("renders without crashing", () => {
+  const wrapper = shallow(<Map />);
+  expect(wrapper).toHaveLength(1);
+});
