+++ src/components/map/index.js
@@ -1,0 +1,3 @@
+import Map from "./Map";
+
+export default Map;
